# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iqrah Backend API",
    "description": "REST API for the Iqrah mobile app",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/admin/packs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Lists all packs and their latest versions for admin tooling.",
        "operationId": "list_all_packs",
        "responses": {
          "200": {
            "description": "All packs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackManifestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Registers a new pack.",
        "operationId": "register_pack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pack registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterPackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/users/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Returns current user profile from JWT context.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "API error response format",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "example": "[\"name: validation failed\"]"
          },
          "error": {
            "type": "string",
            "example": "Validation error: field is required"
          }
        }
      },
...
      "PackId": {
        "type": "string",
        "description": "Stable pack identifier.",
        "example": "translation.en"
      },
      "PackManifestEntry": {
        "type": "object",
        "description": "Global manifest entry for a published active pack.",
        "required": [
          "id",
          "name",
          "pack_type",
          "version",
          "sha256",
          "file_size_bytes",
          "created_at",
          "download_url"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T16:00:00Z"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "download_url": {
            "type": "string"
          },
          "file_size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "$ref": "#/components/schemas/PackId"
          },
          "name": {
            "type": "string"
          },
          "pack_type": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "PackManifestResponse": {
        "type": "object",
        "description": "Global pack manifest response.",
        "required": [
          "packs"
        ],
        "properties": {
          "packs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackManifestEntry"
            }
          }
        }
      },
...
      "PackType": {
        "type": "string",
        "description": "Pack type (translation, recitation, etc.)",
        "enum": [
          "translation",
          "recitation",
          "tafsir"
        ],
        "example": "translation"
      },
...
      "RegisterPackRequest": {
        "type": "object",
        "description": "Request to register a new pack.",
        "required": [
          "name",
          "description",
          "pack_type"
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "English translation pack for mobile clients"
          },
          "name": {
            "type": "string",
            "example": "English Translation"
          },
          "pack_type": {
            "$ref": "#/components/schemas/PackType"
          }
        }
      },
      "RegisterPackResponse": {
        "type": "object",
        "description": "Response for pack registration.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
...
      "UserId": {
        "type": "string",
        "description": "Stable user identifier.",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
      "UserProfile": {
        "type": "object",
        "description": "User profile response.",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T16:00:00Z"
          },
          "id": {
            "$ref": "#/components/schemas/UserId"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "example": "2026-02-20T16:10:00Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}